streamlined-adapter/requirements.lock.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
boto3==1.40.49
botocore==1.40.49
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
curl_cffi==0.13.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
exceptiongroup==1.3.0
fastapi==0.118.2
filelock==3.19.1
Flask==3.1.2
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.79
langchain-text-splitters==0.3.11
langsmith==0.4.34
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
multitasking==0.0.12
networkx==3.2.1
numpy==1.26.4
openai==2.3.0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
peewee==3.18.2
pillow==11.3.0
platformdirs==4.4.0
propcache==0.4.1
protobuf==6.33.0
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
pymongo==4.15.3
python-a2a==0.5.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.2.2
tqdm==4.67.1
transformers==4.57.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.37.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.22.0
yfinance==0.2.66
zipp==3.23.0
zstandard==0.25.0
